feat: Display error cause in RandomGameWidget
This commit updates the `RandomGameWidgetContentFailed` to display the specific error message when fetching game data fails. The error message is passed from `RandomGameWidgetProvider` to the content composable.

Additionally, the `kotlinxSerializationJson` library is updated to version 1.8.1.
The UI configuration for `HomeBannerHorizontalPager` and `HomeRoute` is also adjusted to use the medium layout for expanded window sizes.